Do not print garbage latency if lat_data_get fails to get latency
It is possible for lat_data_get to not succeed in reading the latency,
in which case the latency contains a garbage value and this misleading
value is printed.  Instead, print "unknown" if we cannot fetch the latency
to indicate that it could not be read rather than emiting an incorrect
garbage value.

Signed-off-by: Colin Ian King <colin.king@canonical.com>