Powerpc: Don't show kernel callchain
numatop is not intelligent enough to resolve kernel address to
symbol name and thus it just shows kernel addresses as hex numbers
in call-chain. The real usecase of numatop is to debug numa related
bottlenecks in userspace applications and not the kernel. No need
to show kernel addresses for PowerPC as well.

Signed-off-by: Ravi Bangoria <ravi.bangoria@linux.vnet.ibm.com>