x86: Fix clock frequency parsing
AMD processors do not advertise a base clock frequency as
a part of the "model name" in /proc/cpuinfo. The parsing
must fail in order to let os_calibrate() determine clock
speed from cpufreq information or from TSC instead.

Since the parser fails to find "@", strcspn() returns the
length of the line instead and sscanf() ends up scanning
garbage values beyond the null terminator that match the
format specifier. To avoid this, add an additional check
that makes the condition fail if "@" is not found.

Fixes: eaeed92 ("Powerpc: Fix CPU% utilization for PowerVMs")
Signed-off-by: Sandipan Das <sandipan.das@amd.com>