x86/zen: Add Zen 4 support
Add vendor and family identification as well as the relevant
events to count per-process memory accesseses and CPU usage
on AMD Zen 4 family of processors.

Key changes include the use of the LsNotHaltedP0Cyc event to
count cycles at P0 frequency. This improves the accuracy of
the utilization metrics as, unlike the typical cycles event,
this is clock frequency invariant.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>