common: Add exclude guest to platform event config
Precise PMU events are currently used for capturing memory
access statistics. Currently, these events are programmed to
exclude guests and this does not work well on all platforms
due to differences in the design of the precise PMU.

E.g. precise events are programmed through Instruction Based
Sampling (IBS) on AMD processors but that PMU does not have
the ability to ignore guests unlike the Core PMU.

Add exclude guest as an additional attribute to the platform
event config structure so that precise events can be customized
to work on different platforms.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>